requirements.txt
requests==2.20
pyyaml
pandas